refactor: Rename `testRule` to `rbacRules` and move RBAC tests to their own class
