+++ back/kjar/src/main/resources/rules/rbac/rbacRules.drl
